--- src/comonent/Header.tsx
@@ -1,15 +1,0 @@
-
-function Header() {
-  return (
-    <div>
-        <header className="header">
-        <h2 className="header-title">Submit a Complaint</h2>
-        <p className="header-description">
-          We value your feedback. Please submit your concerns below.
-        </p>
-      </header>
-    </div>
-  )
-}
-
-export default Header
