--- src/comonent/Pagination.tsx
@@ -1,32 +1,0 @@
-type Props = {
-    currentPage: number;
-    totalPages: number;
-    onPageChange: (page: number) => void;
-  };
-  
-  function Pagination({ currentPage, totalPages, onPageChange }: Props) {
-    return (
-      <div className="pagination">
-        <button
-          disabled={currentPage === 1}
-          onClick={() => onPageChange(currentPage - 1)}
-          className="pagination-btn"
-        >
-          Previous
-        </button>
-        <span className="pagination-info">
-          Page {currentPage} of {totalPages}
-        </span>
-        <button
-          disabled={currentPage === totalPages}
-          onClick={() => onPageChange(currentPage + 1)}
-          className="pagination-btn"
-        >
-          Next
-        </button>
-      </div>
-    );
-  }
-  
-  export default Pagination;
-
