--- src/utils/errorHandler.ts
@@ -1,8 +1,0 @@
-export function handleError(error: unknown): string {
-  if (error instanceof Error) return error.message;
-  if (typeof error === "string") return error;
-  if (typeof error === "object" && error !== null && "message" in error) {
-    return String(error.message);
-  }
-  return "An unknown error occurred";
-}
