--- src/utils/wait.ts
@@ -1,3 +1,0 @@
-export const wait = (ms: number): Promise<void> => {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-};
